+++ js/abstract/file_web.js
@@ -1,3 +1,20 @@
-export function fetchVirtualFS(module) {
-    return module.FS;
+import * as wasm from "../wasm.js";
+
+export function writeFile(path, buffer) {
+    wasm.call(module => module.FS.writeFile(path, buffer));
+    return;
+}
+
+export function readFile(path) {
+    wasm.call(module => module.FS.readFile(path, { encoding: 'binary' }));
+    return;
+}
+
+export function removeFile(path) {
+    wasm.call(module => module.FS.unlink(path));
+    return;
+}
+
+export function existsFile(path) {
+    return wasm.call(module => module.FS.analyzePath(path).exists);
 }