+++ js/abstract/realize_web.js
@@ -1,19 +1,13 @@
-import * as wasm from "../wasm.js";
+import * as methods from "../file.js";
 
-export function realize(file, { prefix = "temp_", extension = "" } = {}) {
-    if (typeof file == "string") {
-        throw new Error("cannot access the file system in a browser context");
-    }
-
-    // Choosing a decent path on the VFS.
+export function temp(prefix, extension) {
     let tmppath;
     do {
         tmppath = prefix + String(Number(new Date())) + "_" + String(Math.round(Math.random() * 10000)) + extension;
-    } while (wasm.existsVirtualFile(tmppath));
+    } while (methods.existsFile(tmppath));
+    return tmppath;
+}
 
-    wasm.writeVirtualFile(tmppath, file);
-    return { 
-        path: tmppath, 
-        flush: () => wasm.removeVirtualFile(tmppath) 
-    };
+export function handleString(file) {
+    throw new Error("cannot access the file system in a browser context");
 }